requirements.txt
Django==2.1
django-phone-field==1.8.0
pytz==2019.3
